+++ test/prototypes/test_logistic_request_container.py
@@ -17,6 +17,7 @@ from draftsman.data import mods
 
 from collections.abc import Hashable
 import pytest
+import warnings
 
 
 @pytest.fixture
@@ -420,7 +421,9 @@ class TestRequestContainer:
             "name": "logistic-chest-requester",
             "position": {"x": 0.5, "y": 0.5},
         }
-        chest = LogisticRequestContainer.from_dict(old_dict, version=(1, 0))
+        with warnings.catch_warnings():
+            warnings.filterwarnings("ignore", category=UnknownEntityWarning)
+            chest = LogisticRequestContainer.from_dict(old_dict, version=(1, 0))
         assert chest.to_dict(version=(1, 0)) == old_dict
 
         old_dict_with_filters = {
@@ -434,9 +437,11 @@ class TestRequestContainer:
                 }
             ],
         }
-        chest = LogisticRequestContainer.from_dict(
-            old_dict_with_filters, version=(1, 0)
-        )
+        with warnings.catch_warnings():
+            warnings.filterwarnings("ignore", category=UnknownEntityWarning)
+            chest = LogisticRequestContainer.from_dict(
+                old_dict_with_filters, version=(1, 0)
+            )
         assert len(chest.sections) == 1
         assert chest.sections == [
             ManualSection(